requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-js-asset==1.2.2
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
